requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloud-sql-python-connector @ git+https://github.com/GoogleCloudPlatform/cloud-sql-python-connector@a6fbaee3108e909485ec07697c67c0d140ac3fc5
cryptography==40.0.2
Flask==2.3.2
Flask-Login==0.6.2
frozenlist==1.3.3
google-auth==2.19.0
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
pg8000==1.29.5
praw==7.7.0
prawcore==2.3.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyMySQL==1.0.3
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
scramp==1.4.4
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.2
update-checker==0.18.0
urllib3==1.26.16
websocket-client==1.5.2
Werkzeug==2.3.4
yarl==1.9.2
